Be feature/live auction crud

See merge request s08-webmobile1-sub2/S08P12A704!44

+++ backend/src/main/java/com/auctopus/project/api/service/AuctionImageServiceImpl.java
@@ -13,7 +13,7 @@ public class AuctionImageServiceImpl implements AuctionImageService {
     private AuctionImageRepository auctionImageRepository;
 
     @Override
-    public List<AuctionImage> getAuctionImageListByAuctionSeq(Long auctionSeq) {
+    public List<AuctionImage> getAuctionImageListByAuctionSeq(int auctionSeq) {
         List<AuctionImage> auctionImageList = auctionImageRepository.findByAuctionSeq(auctionSeq)
                 .orElse(null);
         return auctionImageList;
